Paper, Folder, Garbage Can, Couch
I did little research about importing Maya files into Unity. One said
that leaving it as an [.mb] file was okay, so I’m going to leave these
unformatted so we can test out the comparisons/differences. The paper,
folder, and garbage can (hopefully!) have the textures still attached.
The couch, which might change in a future sprint, is unrendered. Let me
know if the files don’t work or if I need to edit anything!